+++ activestorage/test/test_helper.rb
@@ -128,6 +128,7 @@ class User < ActiveRecord::Base
   validates :name, presence: true
 
   has_one_attached :avatar
+  has_one_attached :icon, dependent: :purge
   has_one_attached :cover_photo, dependent: false, service: :local
   has_one_attached :avatar_with_variants do |attachable|
     attachable.variant :thumb, resize_to_limit: [100, 100]
@@ -145,6 +146,7 @@ class User < ActiveRecord::Base
   has_one_attached :name_pronunciation_audio
 
   has_many_attached :highlights
+  has_many_attached :favorites, dependent: :purge
   has_many_attached :vlogs, dependent: false, service: :local
   has_many_attached :highlights_with_variants do |attachable|
     attachable.variant :thumb, resize_to_limit: [100, 100]